requirements.txt
streamlit==1.28.1
pandas==2.1.1
plotly==5.17.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv
psycopg2-binary 